refactored project as a complete module, still depends on robot-hat at the moment
