+++ .gitignore
@@ -128,3 +128,8 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+
+# custom
+old/
+.vscode/
