--- RobotSystems.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		},
-		{
-			"path": "../robot-hat"
-		}
-	],
-	"settings": {}
-}
