--- picar-x/MANIFEST.in
@@ -1,7 +1,0 @@
-include DESCRIPTION.rst
-
-# Include the test suite (FIXME: does not work yet)
-# recursive-include tests *
-
-# If using Python 3.5 or less, then have to include package data, even though
-# it's already declared in setup.py
