--- picar-x/README.md
@@ -1,76 +1,0 @@
-# Picar-X
-
-Picar-X Python library for Raspberry Pi.
-
-Quick Links:
-
-- [Picar-X](#Picar-X)
-  - [About Picar-X](#about-Picar-X)
-  - [Update](#update)
-  - [Installation](#installation)
-  - [Trouble Shooting](#trouble-shooting)
-  - [About SunFounder](#about-sunfounder)
-  - [License](#license)
-  - [Contact us](#contact-us)
-
-## About Robot Hat
-
-TODO
-
-## Update
-
-2021-07-05:
-
-- New Release
-
-## Installation
-
-```bash
-git clone -b robot_hat https://github.com/sunfounder/picar-x.git
-cd picar-x
-sudo python3 setup.py install
-
-# i2c
-
-# Install espeak
-
-sudo apt install espeak
-```
-
-## Trouble Shooting
-
-----------------------------------------------
-
-## About SunFounder
-
-SunFounder is a technology company focused on Raspberry Pi and Arduino open source community development. Committed to the promotion of open source culture, we strives to bring the fun of electronics making to people all around the world and enable everyone to be a maker. Our products include learning kits, development boards, robots, sensor modules and development tools. In addition to high quality products, SunFounder also offers video tutorials to help you make your own project. If you have interest in open source or making something cool, welcome to join us!
-
-----------------------------------------------
-
-## License
-
-This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied wa rranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License along with this program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-
-{Repository Name} comes with ABSOLUTELY NO WARRANTY; for details run ./show w. This is free software, and you are welcome to redistribute it under certain conditions; run ./show c for details.
-
-SunFounder, Inc., hereby disclaims all copyright interest in the program '{Repository Name}' (which makes passes at compilers).
-
-Mike Huang, 21 August 2015
-
-Mike Huang, Chief Executive Officer
-
-Email: service@sunfounder.com, support@sunfounder.com
-
-----------------------------------------------
-
-## Contact us
-
-website:
-    www.sunfounder.com
-
-E-mail:
-    service@sunfounder.com, support@sunfounder.com
