--- picar-x/picarx/sensor.py
@@ -1,49 +1,0 @@
-#!/usr/bin/env/python3
-"""
-sensors.py
-Luke Strohbehn
-"""
-from basic import _Basic_class
-from motor import Motor
-
-
-class Sensor(_Basic_class):
-    def __init__(self) -> None:
-        super().__init__()
-        # ADC
-        self.ADDR = 0x14
-
-    class ADC(Motor.I2C):
-        def __init__(self, chn):  # 参数，通道数，树莓派扩展板上有8个adc通道分别为"A0, A1, A2, A3, A4, A5, A6, A7"
-            super().__init__()
-            if isinstance(chn, str):
-                if chn.startswith("A"):  # 判断穿境来的参数是否为A开头，如果是，取A后面的数字出来
-                    chn = int(chn[1:])
-                else:
-                    raise ValueError("ADC channel should be between [A0, A7], not {0}".format(chn))
-            if chn < 0 or chn > 7:  # 判断取出来的数字是否在0~7的范围内
-                self._error("Incorrect channel range")
-            chn = 7 - chn
-            self.chn = chn | 0x10  # 给从机地址
-            self.reg = 0x40 + self.chn
-            # self.bus = smbus.SMBus(1)
-
-        def read(self):  # adc通道读取数---写一次数据，读取两次数据 （读取的数据范围是0~4095）
-            self._debug("Write 0x%02X to 0x%02X" % (self.chn, self.ADDR))
-            # self.bus.write_byte(self.ADDR, self.chn)      # 写入数据
-            self.send([self.chn, 0, 0], self.ADDR)
-
-            self._debug("Read from 0x%02X" % (self.ADDR))
-            # value_h = self.bus.read_byte(self.ADDR)
-            value_h = self.recv(1, self.ADDR)[0]  # 读取数据
-
-            self._debug("Read from 0x%02X" % (self.ADDR))
-            # value_l = self.bus.read_byte(self.ADDR)
-            value_l = self.recv(1, self.ADDR)[0]  # 读取数据（读两次）
-
-            value = (value_h << 8) + value_l
-            self._debug("Read value: %s" % value)
-            return value
-
-        def read_voltage(self):  # 将读取的数据转化为电压值（0~3.3V）
-            return self.read * 3.3 / 4095
