+++ picarx/basic.py
@@ -1,0 +1,61 @@
+# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
+import logging
+import time
+
+
+class _Basic_class(object):
+    _class_name = "_Basic_class"
+    DEBUG_LEVELS = {
+        "debug": logging.DEBUG,
+        "info": logging.INFO,
+        "warning": logging.WARNING,
+        "error": logging.ERROR,
+        "critical": logging.CRITICAL,
+    }
+    DEBUG_NAMES = ["critical", "error", "warning", "info", "debug"]
+
+    def __init__(self):
+        self._debug_level = 0
+        self.logger = logging.getLogger(self._class_name)
+        self.ch = logging.StreamHandler()
+        form = "%(asctime)s	[%(levelname)s]	%(message)s"
+        self.formatter = logging.Formatter(form)
+        self.ch.setFormatter(self.formatter)
+        self.logger.addHandler(self.ch)
+        self._debug = self.logger.debug
+        self._info = self.logger.info
+        self._warning = self.logger.warning
+        self._error = self.logger.error
+        self._critical = self.logger.critical
+
+    @property
+    def debug(self):
+        return self._debug_level
+
+    @debug.setter
+    def debug(self, debug):
+        if debug in range(5):
+            self._debug_level = self.DEBUG_NAMES[debug]
+        elif debug in self.DEBUG_NAMES:
+            self._debug_level = debug
+        else:
+            raise ValueError(
+                'Debug value must be 0(critical), 1(error), 2(warning), 3(info) or 4(debug), not "{0}".'.format(debug)
+            )
+        self.logger.setLevel(self.DEBUG_LEVELS[self._debug_level])
+        self.ch.setLevel(self.DEBUG_LEVELS[self._debug_level])
+        self._debug("Set logging level to [%s]" % self._debug_level)
+
+    def run_command(self, cmd):
+        import subprocess
+
+        p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+        result = p.stdout.read().decode("utf-8")
+        status = p.poll()
+        # print(result)
+        # print(status)
+        return status, result
+
+    def map(self, x, in_min, in_max, out_min, out_max):
+        return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min
