--- tests/k_turn.py
@@ -1,16 +1,0 @@
-#!/usr/bin/env/python3
-"""
-k_turn.py
-Luke Strohbehn
-"""
-
-import picarx as pcx
-
-
-
-def main():
-    return
-
-
-if __name__ == "__main__":
-    main()
