+++ tests/move_forward.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env/python3
+"""
+move_forward.py
+"""
+from picarx import picarx
+
+
+def main():
+    px=picarx.Picarx()
+    px.forward(40)
+
+
+
+
+    return
+
+if __name__ == "__main__":
+    main()
