--- tests/parallel_park.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env/python3
-"""
-parallel_parking.py
-Luke Strohbehn
-"""
-
-import picarx as pcx
-
-def main():
-    return
-
-
-if __name__ == "__main__":
-    main()
